requirements.txt
openai
pillow
streamlit
fpdf
requests
matplotlib